+++ CHANGELOG.md
@@ -5,6 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [3.9.0] - 2024-08-01
+
+### Changed
+- Fixed Deferred generator block entities using the wrong config-based conditional statement.
+- Updated treasure2-chests-1.20.1 toml file to v3.
+
+### Added
+- Added Deferred generators for Surface, Subaquatic structures, and Pits.
+
+
 ## [3.8.6] - 2024-07-24
 
 ### Changed